--- src/componentes/Tabla/Data_table.css
@@ -1,4 +1,0 @@
-.colorh2{
-    background-color: blue;
-    margin-bottom: 0px;
-}
