--- src/componentes/Tabla/Data_table.js
@@ -1,63 +1,0 @@
-import * as React from 'react';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
-import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
-import Paper from '@mui/material/Paper';
-import { Button } from '@mui/material';
-function createData(name, calories, fat, carbs, protein) {
-  return { name, calories, fat, carbs, protein };
-}
-
-const rows = [
-  createData('Frozen yoghurt', 159, 6.0, 24, 4.0),
-  createData('Ice cream sandwich', 237, 9.0, 37, 4.3),
-  createData('Eclair', 262, 16.0, 24, 6.0),
-  createData('Cupcake', 305, 3.7, 67, 4.3),
-  createData('Gingerbread', 356, 16.0, 49, 3.9),
-];
-
-export default function Data_table() {
-  return (
-    <TableContainer component={Paper}>
-      <Table sx={{ minWidth: 650 }} aria-label="simple table">
-        <TableHead>
-      
-        <h2 style={{ color: '#2074d4' }}>Recent Orders</h2>
-
-          <TableRow>
-          
-            <TableCell>Date </TableCell>
-            <TableCell align="right">Name</TableCell>
-            <TableCell align="right">Ship To</TableCell>
-            <TableCell align="right">Payment Method</TableCell>
-            <TableCell align="right">Sale Amount</TableCell>
-            <TableCell align="right"> <Button variant="contained" color="primary"> Denegar</Button><Button variant="contained" color="primary"> Aceptar</Button></TableCell>
-           
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {rows.map((row) => (
-            <TableRow
-              key={row.name}
-              sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
-            >
-               
-              <TableCell component="th" scope="row">
-                {row.name}
-              </TableCell>
-              <TableCell align="right">{row.calories}</TableCell>
-              <TableCell align="right">{row.fat}</TableCell>
-              <TableCell align="right">{row.carbs}</TableCell>
-              <TableCell align="right">{row.protein}</TableCell>
-    
-            </TableRow>
-          ))}
-           <Button variant="contained" color="primary"> See more orders</Button>
-        </TableBody>
-      </Table>
-    </TableContainer>
-  );
-}
